Abstract out 'PushGH' dialog popup
